requirements.txt
numpy==2.1.1
matplotlib==3.10
Pillow==10.0.0
